pyproject.toml
[tool.poetry]
name = "kurs-7-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"

django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.0"
drf-yasg = "^1.21.7"
setuptools = "^72.1.0"
stripe = "^10.5.0"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.6.0"
eventlet = "^0.36.1"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
